requirements.txt
numpy==2.3.1
pyyaml==6.0.2
fastui==0.7.0
python-multipart==0.0.20